Update config layout
