+++ configs/neovim/config/neovim/init.lua
@@ -1,9 +1,9 @@
-if require('personal.first_load')() then return end
+if require('sv.first_load')() then return end
 
 vim.g.mapleader = ' '
 vim.g.maplocalleader = ' '
 
-require("personal.plugins")
+require("sv.plugins")
 
 vim.lsp.set_log_level("debug")
