+++ configs/neovim/config/neovim/lua/sv/configs/nerdtree.lua
