+++ configs/neovim/config/neovim/lua/sv/configs/quickscope.lua
