+++ configs/neovim/config/neovim/lua/sv/configs/telescope.lua
